Tests/LibWeb/Text/expected/css/cssFloat-property.txt

spanRule: [object CSSStyleRule] ~ span { float: right; }
spanRule.style.cssFloat: right
